User DAO and fixes / New Errors
Login broke for some reason, and I cannot fix it right now.

Whenever you sign in it will act as if your password isn't strong enough or say something along the lines of "something went wrong. please try again later".

Either way, it isn't working, and I've spent nearly an hour trying to fix it. No amount of new emails, password strength lowering, or even resetting the app is working. I'm sure it's something I've done on the app side of things, but I'm curious because there doesn't seem to be much that has changed in that regard.