Working on User Profile
Need to fix up user profile. The main attraction will be the movies piece. I'm not sure how user picture or name can be found. I might have to access oid.

Currently, I need to make a way to get a single movie from tmdb to form an array of user movies for the front end.