Added some profile changes
Working on profile and Auth
req.oidc.user has what I want for the profile, so the search function is back on. I need to add that info to the database, so the user can have a picture and name

Consider changing auth0 registration to include name

Largest issue at the moment is authentication from the front end. Backend authentication is doing fine.